+++ webpack.config.js
@@ -1,4 +1,5 @@
 const path = require('path');
+const webpack = require('webpack');
 const WebpackDashDynamicImport = require('@plotly/webpack-dash-dynamic-import');
 
 const packagejson = require('./package.json');
@@ -62,7 +63,10 @@ module.exports = {
         }
     },
     plugins: [
-        new WebpackDashDynamicImport()
+        new WebpackDashDynamicImport(),
+        new webpack.SourceMapDevToolPlugin({
+            filename: '[file].map'
+        })
     ],
     resolve: {
         alias: {